requirements.txt
svcs 
edgedb 
fastapi 
httpx[cli] 
uvicorn
pytest
pytest-asyncio
pytest-mock
pytest-faker
pydantic-settings
fastui
python-multipart
tzdata 
starlette-csrf
pytest-cov
structlog
asgi-correlation-id
rich
colorama